+++ src/containers/Tenant/Diagnostics/TenantOverview/TenantStorage/TenantStorageNew.scss
@@ -1,5 +1,3 @@
-@use '../../../../../styles/mixins.scss';
-
 .ydb-tenant-storage-new {
     --ydb-storage-segment-row-tables: #5282ff;
     --ydb-storage-segment-column-tables: #ff7112;
@@ -16,278 +14,5 @@
         display: flex;
         flex-direction: column;
         gap: var(--g-spacing-3);
-    }
-
-    &__summary-card {
-        display: flex;
-        flex-direction: column;
-        gap: var(--g-spacing-4);
-
-        padding: var(--g-spacing-4);
-
-        border-radius: var(--g-border-radius-xs);
-        background: var(--g-color-base-generic);
-
-        &_first {
-            border-radius: var(--g-border-radius-m) var(--g-border-radius-xs)
-                var(--g-border-radius-xs) var(--g-border-radius-xs);
-        }
-
-        &_last {
-            border-radius: var(--g-border-radius-xs) var(--g-border-radius-xs)
-                var(--g-border-radius-m) var(--g-border-radius-m);
-        }
-    }
-
-    &__summary-copy {
-        display: flex;
-        flex-direction: column;
-        gap: var(--g-spacing-half);
-
-        min-width: 0;
-    }
-
-    &__summary-metrics {
-        display: flex;
-        align-items: stretch;
-        gap: var(--g-spacing-6);
-    }
-
-    &__summary-metric {
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-
-        min-width: 0;
-        padding-left: var(--g-spacing-6);
-
-        border-left: 1px solid var(--g-color-line-generic);
-
-        &_emphasize {
-            padding: var(--g-spacing-2) var(--g-spacing-3);
-
-            border-left: 0;
-            border-radius: var(--g-border-radius-m);
-            background: var(--g-color-base-generic-accent);
-        }
-
-        &_hide-divider {
-            padding-left: 0;
-
-            border-left: 0;
-        }
-    }
-
-    &__summary-metric-label {
-        margin-top: var(--g-spacing-1);
-    }
-
-    &__summary-progress {
-        width: 100%;
-    }
-
-    &__summary-progress-bar {
-        --g-progress-empty-background-color: var(--g-color-sfx-shadow);
-
-        overflow: hidden;
-
-        height: 20px;
-
-        border-radius: var(--g-border-radius-xs);
-    }
-
-    &__summary-progress-bar .g-progress__item {
-        border-radius: var(--g-border-radius-xs);
-
-        transition: none;
-    }
-
-    &__segment-progress {
-        display: flex;
-        overflow: hidden;
-        gap: 4px;
-
-        height: 20px;
-
-        border-radius: var(--g-border-radius-xs);
-    }
-
-    &__segment-item {
-        min-width: 8px;
-        height: 100%;
-
-        border-radius: var(--g-border-radius-xs);
-
-        &_inactive {
-            opacity: 0.5;
-        }
-
-        &:focus-visible {
-            outline: 2px solid var(--g-color-line-focus);
-            outline-offset: 1px;
-        }
-    }
-
-    &__segment-empty {
-        flex: 1;
-
-        height: 100%;
-
-        opacity: 0.75;
-        border-radius: var(--g-border-radius-xs);
-        background: var(--g-color-sfx-shadow);
-
-        &_inactive {
-            opacity: 0.5;
-        }
-    }
-
-    &__segment-tooltip {
-        margin: 0;
-        padding-inline-start: var(--g-spacing-4);
-
-        white-space: nowrap;
-    }
-
-    &__segment-tooltip li + li {
-        margin-top: var(--g-spacing-half);
-    }
-
-    &__segment-popup {
-        max-width: 360px;
-        padding: var(--g-spacing-1) var(--g-spacing-2);
-
-        color: var(--g-color-text-primary);
-        border-radius: 4px;
-        background: var(--g-color-base-float);
-        box-shadow: 0 1px 5px 0 var(--g-color-sfx-shadow);
-    }
-
-    &__summary-legend {
-        min-width: 0;
-    }
-
-    &__legend-items {
-        display: flex;
-        flex-wrap: wrap;
-        align-items: center;
-        gap: var(--g-spacing-4);
-
-        min-width: 0;
-    }
-
-    &__legend-item {
-        display: flex;
-        align-items: center;
-        gap: var(--g-spacing-2);
-
-        white-space: nowrap;
-
-        &_inactive {
-            opacity: 0.5;
-        }
-
-        &:focus-visible {
-            outline: 2px solid var(--g-color-line-focus);
-            outline-offset: 2px;
-        }
-    }
-
-    &__legend-label {
-        display: inline-flex;
-        align-items: center;
-        gap: var(--g-spacing-1);
-    }
-
-    &__legend-dot {
-        flex: 0 0 8px;
-
-        width: 8px;
-        height: 8px;
-
-        border-radius: 50%;
-    }
-
-    &__summary-used {
-        white-space: nowrap;
-    }
-
-    &__system-tooltip {
-        min-width: 180px;
-    }
-
-    &__system-tooltip-row {
-        white-space: nowrap;
-    }
-
-    &__type-cell {
-        min-width: 0;
-    }
-
-    &__type-icon {
-        display: inline-flex;
-        flex: 0 0 auto;
-        justify-content: center;
-        align-items: center;
-    }
-
-    &__object-cell {
-        min-width: 0;
-    }
-
-    &__object-link {
-        min-width: 0;
-    }
-
-    &__space-cell {
-        flex-wrap: nowrap;
-
-        min-width: 0;
-    }
-
-    &__space-progress {
-        flex: 0 0 148px;
-    }
-
-    &__space-progress-bar {
-        --g-progress-filled-background-color: var(--g-color-base-positive-medium);
-        --g-progress-empty-background-color: var(--g-color-base-positive-light);
-
-        height: 10px;
-
-        border-radius: var(--g-border-radius-xs);
-    }
-
-    &__space-progress-bar .g-progress__item {
-        border-radius: var(--g-border-radius-xs) 0 0 var(--g-border-radius-xs);
-
-        transition: none;
-    }
-
-    &__space-value {
-        min-width: 42px;
-
-        text-align: right;
-        white-space: nowrap;
-    }
-
-    .ydb-stats-wrapper {
-        gap: var(--g-spacing-3);
-    }
-
-    .ydb-stats-wrapper__header {
-        align-items: baseline;
-    }
-
-    .ydb-resizeable-data-table {
-        width: 100%;
-    }
-
-    .data-table__th {
-        @include mixins.subheader-1-typography();
-    }
-
-    .data-table__row {
-        height: 56px;
     }
 }
