+++ src/containers/Tenant/Diagnostics/TenantOverview/TenantStorage/TenantStorageSummaryCard.scss
@@ -1,0 +1,101 @@
+.ydb-tenant-storage-summary-card {
+    display: flex;
+    flex-direction: column;
+    gap: var(--g-spacing-4);
+
+    padding: var(--g-spacing-4);
+
+    border-radius: var(--g-border-radius-xs);
+    background: var(--g-color-base-generic);
+
+    &_first {
+        border-radius: var(--g-border-radius-m) var(--g-border-radius-xs) var(--g-border-radius-xs)
+            var(--g-border-radius-xs);
+    }
+
+    &_last {
+        border-radius: var(--g-border-radius-xs) var(--g-border-radius-xs) var(--g-border-radius-m)
+            var(--g-border-radius-m);
+    }
+
+    &__copy {
+        display: flex;
+        flex-direction: column;
+        gap: var(--g-spacing-half);
+
+        min-width: 0;
+    }
+
+    &__metrics {
+        display: flex;
+        align-items: stretch;
+        gap: var(--g-spacing-6);
+    }
+
+    &__metric {
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+
+        min-width: 0;
+        padding-left: var(--g-spacing-6);
+
+        border-left: 1px solid var(--g-color-line-generic);
+
+        &_emphasize {
+            padding: var(--g-spacing-2) var(--g-spacing-3);
+
+            border-left: 0;
+            border-radius: var(--g-border-radius-m);
+            background: var(--g-color-base-generic-accent);
+        }
+
+        &_hide-divider {
+            padding-left: 0;
+
+            border-left: 0;
+        }
+    }
+
+    &__metric-label {
+        margin-top: var(--g-spacing-1);
+    }
+
+    &__progress {
+        width: 100%;
+    }
+
+    &__progress-bar {
+        --g-progress-empty-background-color: var(--g-color-sfx-shadow);
+
+        overflow: hidden;
+
+        height: 20px;
+
+        border-radius: var(--g-border-radius-xs);
+    }
+
+    &__progress-bar .g-progress__item {
+        border-radius: var(--g-border-radius-xs);
+
+        transition: none;
+    }
+
+    &__used {
+        white-space: nowrap;
+    }
+
+    &__segment-popup {
+        overflow: hidden;
+
+        box-sizing: border-box;
+        max-width: 360px;
+        max-height: 368px;
+        padding: var(--g-tooltip-padding, var(--g-spacing-1) var(--g-spacing-2));
+
+        color: var(--g-tooltip-text-color, var(--g-color-text-primary));
+        border-radius: var(--g-tooltip-border-radius, 4px);
+        background: var(--g-tooltip-background-color, var(--g-color-base-float));
+        box-shadow: var(--g-tooltip-box-shadow, 0 1px 5px 0 var(--g-color-sfx-shadow));
+    }
+}
