+++ src/containers/Tenant/Diagnostics/TenantOverview/TenantStorage/TenantStorageSummarySections.scss
@@ -1,0 +1,3 @@
+.ydb-tenant-storage-summary-sections {
+    min-width: 0;
+}
